Classes created to calculate taxes
Here i created 3 classes to operate the calculation of taxes. One for the getters and setters for the amount of money earned, taxCredits the person is entitled to deduct and its liquid amount. The class Taxes to declare the 3 different type of taxes due in Ireland and the total amount due, and their getters and setters. The last class, Calculate tax will be to add methods to make the operations.